refactor: import defer registry types directly instead of via re-exports
Internal modules now import DeferRegistry types and the eviction
options from deferRegistry.ts directly; defer.tsx no longer re-exports
them. The public API (defer, DeferOptions) is unchanged.
