+++ packages/static/src/rsc/entry.tsx
@@ -2,7 +2,8 @@ import "./defer";
 import { renderToReadableStream } from "@vitejs/plugin-rsc/rsc";
 import { devMainRscPath } from "./request";
 import { generateAppMarker } from "./marker";
-import { deferRegistry, devDeferEvictionOptions } from "./defer";
+import { deferRegistry } from "./defer";
+import { devDeferEvictionOptions } from "./deferRegistry";
 import { extractIDFromModulePath } from "./rscModule";
 import { stripBasePath } from "../util/basePath";
 import { urlPathToFileCandidates } from "../util/urlPath";
